requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
ccxt==4.4.75
python-dotenv==1.0.1
